Rename from "central" to "center"
Follow the new name of the app: bloomnet.center
