+++ Gemfile
@@ -1,6 +1,6 @@
 source 'https://rubygems.org'
 
-# Specify your gem's dependencies in bloom_net_central_client.gemspec
+# Specify your gem's dependencies in bloom_net_center_client.gemspec
 gemspec
 
 gem "pry-byebug"
