+++ lib/bloom_net_center_client/coercers/indifferent_hash.rb
@@ -1,0 +1,11 @@
+module BloomNetCenterClient
+  module Coercers
+    class IndifferentHash < Virtus::Attribute
+
+      def coerce(value)
+        value.is_a?(::Hash) ? value.with_indifferent_access : value
+      end
+
+    end
+  end
+end
