+++ lib/bloom_net_center_client/models/claim.rb
@@ -1,0 +1,11 @@
+module BloomNetCenterClient
+  class Claim
+
+    include Virtus.model
+    attribute :recipient_first_name, String
+    attribute :recipient_last_name, String
+    attribute :ref_no, String
+    attribute :status, String
+
+  end
+end
