+++ lib/bloom_net_center_client/requests/create_txn_request.rb
@@ -1,0 +1,28 @@
+module BloomNetCenterClient
+  class CreateTxnRequest < BaseRequest
+
+    attribute :recipient_first_name, String
+    attribute :recipient_last_name, String
+    attribute :body, String, lazy: true, default: :default_body
+
+    private
+
+    def default_action
+      :post
+    end
+
+    def path
+      "/api/v1/txns"
+    end
+
+    def default_body
+      {
+        txn: {
+          recipient_first_name: recipient_first_name,
+          recipient_last_name: recipient_last_name,
+        }
+      }.to_json
+    end
+
+  end
+end
