+++ lib/bloom_net_center_client/responses/create_claim_response.rb
@@ -1,0 +1,19 @@
+module BloomNetCenterClient
+  class CreateClaimResponse < BaseResponse
+
+    attribute :claim, Object, lazy: true, default: :default_claim
+
+    private
+
+    def default_claim
+      claim_attrs = body[:data].slice(*%i[
+        recipient_first_name
+        recipient_last_name
+        ref_no
+        status
+      ])
+      BloomNetCenterClient::Claim.new(claim_attrs)
+    end
+
+  end
+end
