+++ lib/bloom_net_center_client/version.rb
@@ -1,0 +1,3 @@
+module BloomNetCenterClient
+  VERSION = "0.2.0"
+end
