--- lib/bloom_net_central_client/coercers/indifferent_hash.rb
@@ -1,11 +1,0 @@
-module BloomNetCentralClient
-  module Coercers
-    class IndifferentHash < Virtus::Attribute
-
-      def coerce(value)
-        value.is_a?(::Hash) ? value.with_indifferent_access : value
-      end
-
-    end
-  end
-end
