--- lib/bloom_net_central_client/requests/base_request.rb
@@ -1,17 +1,0 @@
-module BloomNetCentralClient
-  class BaseRequest
-
-    include APIClientBase::Request.module
-    attribute(:headers, Coercers::IndifferentHash, {
-      lazy: true,
-      default: :default_headers,
-    })
-
-    private
-
-    def default_headers
-      {"Content-Type" => "application/json"}
-    end
-
-  end
-end
