--- lib/bloom_net_central_client/responses/create_claim_response.rb
@@ -1,19 +1,0 @@
-module BloomNetCentralClient
-  class CreateClaimResponse < BaseResponse
-
-    attribute :claim, Object, lazy: true, default: :default_claim
-
-    private
-
-    def default_claim
-      claim_attrs = body[:data].slice(*%i[
-        recipient_first_name
-        recipient_last_name
-        ref_no
-        status
-      ])
-      BloomNetCentralClient::Claim.new(claim_attrs)
-    end
-
-  end
-end
