--- lib/bloom_net_central_client/version.rb
@@ -1,3 +1,0 @@
-module BloomNetCentralClient
-  VERSION = "0.2.0"
-end
