+++ spec/lib/bloom_net_center_client/client_spec.rb
@@ -1,0 +1,7 @@
+require "spec_helper"
+
+module BloomNetCenterClient
+  RSpec.describe Client do
+
+  end
+end
