+++ spec/lib/bloom_net_center_client/requests/base_request_spec.rb
@@ -1,0 +1,12 @@
+require 'spec_helper'
+
+module BloomNetCenterClient
+  RSpec.describe BaseRequest, type: %i[virtus] do
+
+    describe "attributes" do
+      subject { described_class }
+      it { is_expected.to have_attribute(:headers) }
+    end
+
+  end
+end
