+++ spec/lib/bloom_net_center_client/requests/find_txn_request_spec.rb
@@ -1,0 +1,18 @@
+require 'spec_helper'
+
+module BloomNetCenterClient
+  RSpec.describe FindTxnRequest, type: %i[virtus] do
+
+    it "inherits from BaseRequest" do
+      expect(described_class < BaseRequest).to be true
+    end
+
+    describe "attributes" do
+      subject { described_class }
+      it { is_expected.to have_attribute(:ref_no, String) }
+      it { is_expected.to have_attribute(:recipient_first_name, String) }
+      it { is_expected.to have_attribute(:recipient_last_name, String) }
+    end
+
+  end
+end
