--- spec/lib/bloom_net_central_client/client_spec.rb
@@ -1,7 +1,0 @@
-require "spec_helper"
-
-module BloomNetCentralClient
-  RSpec.describe Client do
-
-  end
-end
