--- spec/lib/bloom_net_central_client/requests/base_request_spec.rb
@@ -1,12 +1,0 @@
-require 'spec_helper'
-
-module BloomNetCentralClient
-  RSpec.describe BaseRequest, type: %i[virtus] do
-
-    describe "attributes" do
-      subject { described_class }
-      it { is_expected.to have_attribute(:headers) }
-    end
-
-  end
-end
