+++ spec/spec_helper.rb
@@ -1,4 +1,4 @@
 $LOAD_PATH.unshift File.expand_path("../../lib", __FILE__)
 require "pry-byebug"
 require "virtus-matchers"
-require "bloom_net_central_client"
+require "bloom_net_center_client"
